--- lauthai-spa/src/app/_models/_consts/const.ts
@@ -1,5 +1,0 @@
-export class Const {
-  public static readonly TABLE_COLUMN = [
-    'id', 'name', 'pfpUrl', 'university', 'age', 'job', 'marriedStatus', 'district', 'phone', 'action'
-  ];
-}
