+++ lauthai-spa/src/app/_models/consts/const.ts
@@ -1,0 +1,9 @@
+export class Const {
+  public static readonly TABLE_ADMIN_COLUMN = [
+    'id', 'name', 'pfpUrl', 'university', 'age', 'job', 'marriedStatus', 'district', 'phone', 'action'
+  ];
+
+  public static readonly TABLE_USER_COLUMN = [
+    'id', 'name', 'pfpUrl', 'university', 'age', 'job', 'marriedStatus', 'district', 'phone',
+  ];
+}
