+++ lauthai-spa/src/app/admin/management/create-profile-dialog/create-profile-dialog.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { IProfile } from './../../../_models/_interfaces/profile.interface';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
+
+import { IProfile } from './../../../_models/interfaces/profile.interface';
 
 @Component({
   selector: 'app-admin-create-profile-dialog',
