+++ lauthai-spa/src/app/admin/management/management.component.ts
@@ -1,11 +1,12 @@
 import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import { MatTableDataSource, MatTable } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { IProfile } from './../../_models/_interfaces/profile.interface';
-import { Const } from './../../_models/_consts/const';
-import { ProfileService } from '../../_services/profile.service';
 import { MatDialog } from '@angular/material/dialog';
+import { MatSort } from '@angular/material/sort';
+
+import { IProfile } from './../../_models/interfaces/profile.interface';
+import { Const } from './../../_models/consts/const';
+import { ProfileService } from '../../_services/profile.service';
 import { CreateProfileDialogComponent } from './create-profile-dialog/create-profile-dialog.component';
 import { UpdateProfileDialogComponent } from './update-profile-dialog/update-profile-dialog.component';
 import { DeleteProfileDialogComponent } from './delete-profile-dialog/delete-profile-dialog.component';
@@ -19,7 +20,7 @@ import { DeleteProfileDialogComponent } from './delete-profile-dialog/delete-pro
 
 export class ManagementComponent implements OnInit, AfterViewInit {
 
-  displayedColumns: string[] = Const.TABLE_COLUMN;
+  displayedColumns: string[] = Const.TABLE_ADMIN_COLUMN;
   dataSource: MatTableDataSource<IProfile>;
   profiles: IProfile[];
